1, Add <Library/DevicePathLib.h> for all source that use device path utility macros 2, Add DevicePathLib library class to module's INF that use device path utility macros
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@6456 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/Network/IScsiDxe/IScsiExtScsiPassThru.c
@@ -188,7 +188,7 @@ IScsiExtScsiPassThruBuildDevicePath (
 
   Node->DevPath.Type    = MESSAGING_DEVICE_PATH;
   Node->DevPath.SubType = MSG_ISCSI_DP;
-  SetDevicePathNodeLength (&Node->DevPath, DevPathNodeLen);
+  SetDevicePathNodeLength (&Node->DevPath, (UINT16)DevPathNodeLen);
 
   //
   // 0 for TCP, others are reserved.
